summarize TimeSpatData in Time-Dimension

summary_time_tsd summarize the Time-Dimension in fixed time interval like yearly, monthly and so on

summary_index_tsd summarize the Time-Dimension in flexibel time interval, use the parameter idx_End to divide the interval

summary_time_tsd(tsd_Data, str_Interval = "years", smry_Funct = mean, ...)

summary_index_tsd(tsd_Data, idx_End, smry_Funct = mean, ...)

Arguments

tsd_Data

TimeSpatData data

TimeVectVariable

TimeVectArray

TimeRastVariable

TimeRastArray

str_Interval

(string) <“microseconds”, “milliseconds”, “seconds”, “minutes”, “hours”, “days”, “weeks”, “months”, “quarters” and “years”> time interval

smry_Funct

(function) function that used for summary, e.g. mean, max and min, it can also be the self-defined function

...

other parameters for smry_Funct

idx_End

(integer) index of the end points every interval

Value

summarized time series with the same data structure like tsd_Data